Fix 테스트 자동화 구조 추가
