--- .github/workflows/unit-test.yml
@@ -1,18 +1,0 @@
-name: Unit Test
-
-on:
-  pull_request:
-    branches:
-      - main
-  workflow_dispatch:
-
-jobs:
-  test:
-    runs-on: ubuntu-latest
-
-    steps:
-      - name: Checkout repository
-        uses: actions/checkout@v3
-
-      - name: Run TestNG tests
-        run: ./gradlew clean testMinAndMaxWithXml
