+++ src/main/kotlin/mobile/base/BaseView.kt
@@ -1,0 +1,9 @@
+package mobile.base
+
+import mobile.utilities.ActionUtil
+
+interface BaseView : ActionUtil {
+
+        fun waitUntilOpenView()
+
+}
