+++ src/main/kotlin/mobile/enums/Platform.kt
@@ -1,0 +1,5 @@
+package mobile.enums
+
+enum class Platform() {
+    ANDROID, IOS
+}
