+++ src/test/resources/test-android.xml
@@ -1,0 +1,8 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="AndroidTestSuite" parallel="methods" thread-count="1">
+    <test name="AndroidTest">
+        <classes>
+            <class name="mobile.tests.AndroidTest"/>
+        </classes>
+    </test>
+</suite>
