+++ src/test/resources/test-ios.xml
@@ -1,0 +1,8 @@
+<!DOCTYPE suite SYSTEM "https://testng.org/testng-1.0.dtd">
+<suite name="IOSTestSuite" parallel="methods" thread-count="1">
+    <test name="IOSTest">
+        <classes>
+            <class name="mobile.tests.IOSTest"/>
+        </classes>
+    </test>
+</suite>
